+++ src/test/java/com/newbit/coffeeletter/service/ChatServiceTest.java
@@ -2,78 +2,60 @@ package com.newbit.coffeeletter.service;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
-import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 
 import com.newbit.coffeeletter.domain.chat.ChatMessage;
 import com.newbit.coffeeletter.domain.chat.CoffeeLetterRoom;
 import com.newbit.coffeeletter.domain.chat.MessageType;
 import com.newbit.coffeeletter.dto.ChatMessageDTO;
 import com.newbit.coffeeletter.dto.CoffeeLetterRoomDTO;
-import com.newbit.coffeeletter.repository.ChatMessageRepository;
-import com.newbit.coffeeletter.repository.CoffeeLetterRoomRepository;
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
 class ChatServiceTest {
 
     @Mock
-    private CoffeeLetterRoomRepository roomRepository;
+    private RoomService roomService;
     
     @Mock
-    private ChatMessageRepository messageRepository;
-    
-    @Mock
-    private SimpMessagingTemplate messagingTemplate;
-
-    @Mock
-    private ModelMapper modelMapper;
+    private MessageService messageService;
 
     @InjectMocks
     private ChatServiceImpl chatService;
 
     private CoffeeLetterRoomDTO roomDTO;
     private CoffeeLetterRoom room;
-    private List<CoffeeLetterRoom> rooms;
+    private List<CoffeeLetterRoomDTO> roomDTOs;
     private ChatMessageDTO messageDTO;
     private ChatMessage message;
-    private List<ChatMessage> messages;
-    private Page<ChatMessage> messagePage;
+    private List<ChatMessageDTO> messageDTOs;
+    private Page<ChatMessageDTO> messagePage;
 
     @BeforeEach
     void setUp() {
         roomDTO = new CoffeeLetterRoomDTO();
+        roomDTO.setId("test-room-id");
         roomDTO.setMentorId(1L);
         roomDTO.setMenteeId(2L);
         roomDTO.setMentorName("멘토");
@@ -91,14 +73,16 @@ class ChatServiceTest {
         room.getParticipants().add("1");
         room.getParticipants().add("2");
 
-        rooms = Arrays.asList(room);
+        roomDTOs = Arrays.asList(roomDTO);
         
         messageDTO = new ChatMessageDTO();
+        messageDTO.setId("test-message-id");
         messageDTO.setRoomId("test-room-id");
         messageDTO.setSenderId(1L);
         messageDTO.setSenderName("멘토");
         messageDTO.setContent("안녕하세요");
         messageDTO.setType(MessageType.CHAT);
+        messageDTO.setTimestamp(LocalDateTime.now());
         
         message = new ChatMessage();
         message.setId("test-message-id");
@@ -112,23 +96,16 @@ class ChatServiceTest {
         message.setReadByMentee(false);
         
         // 메시지 목록 및 페이지 설정
-        messages = Arrays.asList(message);
-        messagePage = new PageImpl<>(messages);
-        
-        when(modelMapper.map(any(Object.class), eq(CoffeeLetterRoomDTO.class))).thenReturn(roomDTO);
-        when(modelMapper.map(any(Object.class), eq(ChatMessageDTO.class))).thenReturn(messageDTO);
-        when(modelMapper.map(any(CoffeeLetterRoomDTO.class), eq(CoffeeLetterRoom.class))).thenReturn(room);
-        when(modelMapper.map(any(ChatMessageDTO.class), eq(ChatMessage.class))).thenReturn(message);
+        messageDTOs = Arrays.asList(messageDTO);
+        messagePage = new PageImpl<>(messageDTOs);
     }
+
+    // === RoomService 메서드 테스트 ===
 
     @Test
     void createRoom_새로운_채팅방_생성_성공() {
         // given
-        when(roomRepository.findByCoffeeChatId(anyLong()))
-                .thenReturn(Optional.empty());
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(room);
-        when(roomRepository.findById(anyString())).thenReturn(Optional.of(room));
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(message);
+        when(roomService.createRoom(any(CoffeeLetterRoomDTO.class))).thenReturn(roomDTO);
 
         // when
         CoffeeLetterRoomDTO result = chatService.createRoom(roomDTO);
@@ -138,105 +115,41 @@ class ChatServiceTest {
         assertEquals(roomDTO.getMentorId(), result.getMentorId());
         assertEquals(roomDTO.getMenteeId(), result.getMenteeId());
 
-        verify(roomRepository, times(2)).save(any(CoffeeLetterRoom.class));
-
-    }
-
-    @Test
-    void createRoom_이미_존재하는_채팅방_예외발생() {
-        // given
-        when(roomRepository.findByCoffeeChatId(anyLong()))
-                .thenReturn(Optional.of(room));
-
-        // when & then
-        assertThrows(IllegalStateException.class, () -> {
-            chatService.createRoom(roomDTO);
-        });
-        verify(roomRepository, never()).save(any(CoffeeLetterRoom.class));
+        verify(roomService, times(1)).createRoom(roomDTO);
     }
     
     @Test
     void endRoom_채팅방_비활성화_성공() {
         // given
         String roomId = "test-room-id";
-        CoffeeLetterRoom activeRoom = new CoffeeLetterRoom();
-        activeRoom.setId(roomId);
-        activeRoom.setStatus(CoffeeLetterRoom.RoomStatus.ACTIVE);
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(activeRoom));
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(activeRoom);
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(message);
+        when(roomService.endRoom(roomId)).thenReturn(roomDTO);
         
         // when
         CoffeeLetterRoomDTO result = chatService.endRoom(roomId);
         
         // then
         assertNotNull(result);
-        verify(roomRepository, atLeastOnce()).findById(roomId);
-        verify(roomRepository, times(2)).save(any(CoffeeLetterRoom.class));
-        assertEquals(CoffeeLetterRoom.RoomStatus.INACTIVE, activeRoom.getStatus());
-        verify(messageRepository, times(1)).save(any(ChatMessage.class));
-        verify(messagingTemplate, times(1)).convertAndSend(anyString(), any(Object.class));
-    }
-    
-    @Test
-    void endRoom_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.endRoom(roomId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(roomRepository, never()).save(any(CoffeeLetterRoom.class));
+        verify(roomService, times(1)).endRoom(roomId);
     }
     
     @Test
     void cancelRoom_채팅방_취소_성공() {
         // given
         String roomId = "test-room-id";
-        CoffeeLetterRoom activeRoom = new CoffeeLetterRoom();
-        activeRoom.setId(roomId);
-        activeRoom.setStatus(CoffeeLetterRoom.RoomStatus.ACTIVE);
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(activeRoom));
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(activeRoom);
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(message);
+        when(roomService.cancelRoom(roomId)).thenReturn(roomDTO);
         
         // when
         CoffeeLetterRoomDTO result = chatService.cancelRoom(roomId);
         
         // then
         assertNotNull(result);
-        verify(roomRepository, atLeastOnce()).findById(roomId);
-        verify(roomRepository, times(2)).save(any(CoffeeLetterRoom.class));
-        assertEquals(CoffeeLetterRoom.RoomStatus.CANCELED, activeRoom.getStatus());
-        verify(messageRepository, times(1)).save(any(ChatMessage.class));
-        verify(messagingTemplate, times(1)).convertAndSend(anyString(), any(Object.class));
-    }
-    
-    @Test
-    void cancelRoom_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.cancelRoom(roomId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(roomRepository, never()).save(any(CoffeeLetterRoom.class));
+        verify(roomService, times(1)).cancelRoom(roomId);
     }
     
     @Test
     void getAllRooms_모든_채팅방_조회_성공() {
         // given
-        when(roomRepository.findAll()).thenReturn(rooms);
+        when(roomService.getAllRooms()).thenReturn(roomDTOs);
         
         // when
         List<CoffeeLetterRoomDTO> result = chatService.getAllRooms();
@@ -244,28 +157,14 @@ class ChatServiceTest {
         // then
         assertNotNull(result);
         assertEquals(1, result.size());
-        verify(roomRepository, times(1)).findAll();
-    }
-    
-    @Test
-    void getAllRooms_빈_목록_반환() {
-        // given
-        when(roomRepository.findAll()).thenReturn(Collections.emptyList());
-        
-        // when
-        List<CoffeeLetterRoomDTO> result = chatService.getAllRooms();
-        
-        // then
-        assertNotNull(result);
-        assertEquals(0, result.size());
-        verify(roomRepository, times(1)).findAll();
+        verify(roomService, times(1)).getAllRooms();
     }
     
     @Test
     void getRoomById_채팅방_조회_성공() {
         // given
         String roomId = "test-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
+        when(roomService.getRoomById(roomId)).thenReturn(roomDTO);
         
         // when
         CoffeeLetterRoomDTO result = chatService.getRoomById(roomId);
@@ -274,29 +173,14 @@ class ChatServiceTest {
         assertNotNull(result);
         assertEquals(roomDTO.getMentorId(), result.getMentorId());
         assertEquals(roomDTO.getMenteeId(), result.getMenteeId());
-        verify(roomRepository, times(1)).findById(roomId);
-    }
-    
-    @Test
-    void getRoomById_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getRoomById(roomId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
+        verify(roomService, times(1)).getRoomById(roomId);
     }
     
     @Test
     void getRoomsByUserId_사용자_채팅방_조회_성공() {
         // given
         Long userId = 1L;
-        String userIdStr = userId.toString();
-        when(roomRepository.findByParticipantsContaining(userIdStr)).thenReturn(rooms);
+        when(roomService.getRoomsByUserId(userId)).thenReturn(roomDTOs);
         
         // when
         List<CoffeeLetterRoomDTO> result = chatService.getRoomsByUserId(userId);
@@ -304,32 +188,15 @@ class ChatServiceTest {
         // then
         assertNotNull(result);
         assertEquals(1, result.size());
-        verify(roomRepository, times(1)).findByParticipantsContaining(userIdStr);
-    }
-    
-    @Test
-    void getRoomsByUserId_빈_목록_반환() {
-        // given
-        Long userId = 999L;
-        String userIdStr = userId.toString();
-        when(roomRepository.findByParticipantsContaining(userIdStr)).thenReturn(Collections.emptyList());
-        
-        // when
-        List<CoffeeLetterRoomDTO> result = chatService.getRoomsByUserId(userId);
-        
-        // then
-        assertNotNull(result);
-        assertEquals(0, result.size());
-        verify(roomRepository, times(1)).findByParticipantsContaining(userIdStr);
+        verify(roomService, times(1)).getRoomsByUserId(userId);
     }
     
     @Test
     void getRoomsByUserIdAndStatus_사용자와_상태별_채팅방_조회_성공() {
         // given
         Long userId = 1L;
-        String userIdStr = userId.toString();
         CoffeeLetterRoom.RoomStatus status = CoffeeLetterRoom.RoomStatus.ACTIVE;
-        when(roomRepository.findByParticipantsContainingAndStatus(userIdStr, status)).thenReturn(rooms);
+        when(roomService.getRoomsByUserIdAndStatus(userId, status)).thenReturn(roomDTOs);
         
         // when
         List<CoffeeLetterRoomDTO> result = chatService.getRoomsByUserIdAndStatus(userId, status);
@@ -337,80 +204,22 @@ class ChatServiceTest {
         // then
         assertNotNull(result);
         assertEquals(1, result.size());
-        verify(roomRepository, times(1)).findByParticipantsContainingAndStatus(userIdStr, status);
+        verify(roomService, times(1)).getRoomsByUserIdAndStatus(userId, status);
     }
     
-    @Test
-    void getRoomsByUserIdAndStatus_빈_목록_반환() {
-        // given
-        Long userId = 999L;
-        String userIdStr = userId.toString();
-        CoffeeLetterRoom.RoomStatus status = CoffeeLetterRoom.RoomStatus.INACTIVE;
-        when(roomRepository.findByParticipantsContainingAndStatus(userIdStr, status)).thenReturn(Collections.emptyList());
-        
-        // when
-        List<CoffeeLetterRoomDTO> result = chatService.getRoomsByUserIdAndStatus(userId, status);
-        
-        // then
-        assertNotNull(result);
-        assertEquals(0, result.size());
-        verify(roomRepository, times(1)).findByParticipantsContainingAndStatus(userIdStr, status);
-    }
+    // === MessageService 메서드 테스트 ===
     
     @Test
-    void sendMessage_멘토가_보낸_메시지_성공() {
+    void sendMessage_메시지_전송_성공() {
         // given
-        String roomId = "test-room-id";
-        Long mentorId = 1L;
-        
-        messageDTO.setSenderId(mentorId);
-        message.setSenderId(mentorId);
-        message.setReadByMentor(true);
-        message.setReadByMentee(false);
-        
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(message);
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(room);
+        when(messageService.sendMessage(any(ChatMessageDTO.class))).thenReturn(messageDTO);
         
         // when
         ChatMessageDTO result = chatService.sendMessage(messageDTO);
         
         // then
         assertNotNull(result);
-        verify(messageRepository, times(1)).save(any(ChatMessage.class));
-        verify(roomRepository, atLeastOnce()).findById(roomId);
-        verify(roomRepository, times(1)).save(any(CoffeeLetterRoom.class));
-        verify(messagingTemplate, times(1)).convertAndSend(eq("/topic/chat/room/" + roomId), any(Object.class));
-        assertTrue(message.isReadByMentor());
-        assertEquals(false, message.isReadByMentee());
-    }
-    
-    @Test
-    void sendMessage_멘티가_보낸_메시지_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long menteeId = 2L;
-        
-        messageDTO.setSenderId(menteeId);
-        message.setSenderId(menteeId);
-        message.setReadByMentor(false);
-        message.setReadByMentee(true);
-        
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(message);
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(room);
-        
-        // when
-        ChatMessageDTO result = chatService.sendMessage(messageDTO);
-        
-        // then
-        assertNotNull(result);
-        verify(messageRepository, times(1)).save(any(ChatMessage.class));
-        verify(roomRepository, atLeastOnce()).findById(roomId);
-        verify(roomRepository, times(1)).save(any(CoffeeLetterRoom.class));
-        verify(messagingTemplate, times(1)).convertAndSend(eq("/topic/chat/room/" + roomId), any(Object.class));
-        assertTrue(message.isReadByMentee());
-        assertEquals(false, message.isReadByMentor());
+        verify(messageService, times(1)).sendMessage(messageDTO);
     }
     
     @Test
@@ -419,52 +228,21 @@ class ChatServiceTest {
         String roomId = "test-room-id";
         String content = "시스템 메시지 테스트";
         
-        ChatMessage systemMessage = new ChatMessage();
-        systemMessage.setRoomId(roomId);
-        systemMessage.setType(MessageType.SYSTEM);
-        systemMessage.setContent(content);
-        systemMessage.setSenderId(0L);
-        systemMessage.setSenderName("System");
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.save(any(ChatMessage.class))).thenReturn(systemMessage);
-        when(roomRepository.save(any(CoffeeLetterRoom.class))).thenReturn(room);
+        when(messageService.sendSystemMessage(roomId, content)).thenReturn(messageDTO);
         
         // when
         ChatMessageDTO result = chatService.sendSystemMessage(roomId, content);
         
         // then
         assertNotNull(result);
-        verify(roomRepository, atLeastOnce()).findById(roomId);
-        verify(messageRepository, times(1)).save(any(ChatMessage.class));
-        verify(roomRepository, times(1)).save(any(CoffeeLetterRoom.class));
-        verify(messagingTemplate, times(1)).convertAndSend(eq("/topic/chat/room/" + roomId), any(Object.class));
-    }
-    
-    @Test
-    void sendSystemMessage_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        String content = "시스템 메시지 테스트";
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.sendSystemMessage(roomId, content);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).save(any(ChatMessage.class));
-        verify(messagingTemplate, never()).convertAndSend(anyString(), any(Object.class));
+        verify(messageService, times(1)).sendSystemMessage(roomId, content);
     }
     
     @Test
     void getMessagesByRoomId_채팅방_메시지_조회_성공() {
         // given
         String roomId = "test-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomId(roomId)).thenReturn(messages);
+        when(messageService.getMessagesByRoomId(roomId)).thenReturn(messageDTOs);
         
         // when
         List<ChatMessageDTO> result = chatService.getMessagesByRoomId(roomId);
@@ -472,40 +250,7 @@ class ChatServiceTest {
         // then
         assertNotNull(result);
         assertEquals(1, result.size());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomId(roomId);
-    }
-    
-    @Test
-    void getMessagesByRoomId_빈_목록_반환() {
-        // given
-        String roomId = "test-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomId(roomId)).thenReturn(Collections.emptyList());
-        
-        // when
-        List<ChatMessageDTO> result = chatService.getMessagesByRoomId(roomId);
-        
-        // then
-        assertNotNull(result);
-        assertEquals(0, result.size());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomId(roomId);
-    }
-    
-    @Test
-    void getMessagesByRoomId_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getMessagesByRoomId(roomId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomId(anyString());
+        verify(messageService, times(1)).getMessagesByRoomId(roomId);
     }
     
     @Test
@@ -516,8 +261,7 @@ class ChatServiceTest {
         int size = 10;
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "timestamp"));
         
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomId(eq(roomId), any(Pageable.class))).thenReturn(messagePage);
+        when(messageService.getMessagesByRoomId(roomId, pageable)).thenReturn(messagePage);
         
         // when
         Page<ChatMessageDTO> result = chatService.getMessagesByRoomId(roomId, pageable);
@@ -525,437 +269,84 @@ class ChatServiceTest {
         // then
         assertNotNull(result);
         assertEquals(1, result.getContent().size());
-        assertEquals(1, result.getTotalElements());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomId(eq(roomId), any(Pageable.class));
+        verify(messageService, times(1)).getMessagesByRoomId(roomId, pageable);
     }
     
     @Test
-    void getMessagesByRoomId_빈_페이지_반환() {
+    void getUnreadMessages_읽지않은_메시지_조회_성공() {
         // given
         String roomId = "test-room-id";
-        int page = 0;
-        int size = 10;
-        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "timestamp"));
-        Page<ChatMessage> emptyPage = new PageImpl<>(Collections.emptyList());
+        Long userId = 1L;
         
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomId(eq(roomId), any(Pageable.class))).thenReturn(emptyPage);
+        when(messageService.getUnreadMessages(roomId, userId)).thenReturn(messageDTOs);
         
         // when
-        Page<ChatMessageDTO> result = chatService.getMessagesByRoomId(roomId, pageable);
-        
-        // then
-        assertNotNull(result);
-        assertEquals(0, result.getContent().size());
-        assertEquals(0, result.getTotalElements());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomId(eq(roomId), any(Pageable.class));
-    }
-    
-    @Test
-    void getMessagesByRoomId_페이징_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        int page = 0;
-        int size = 10;
-        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "timestamp"));
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getMessagesByRoomId(roomId, pageable);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomId(anyString(), any(Pageable.class));
-    }
-    
-    @Test
-    void getUnreadMessages_멘토의_읽지않은_메시지_조회_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long mentorId = 1L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomIdAndReadByMentorFalse(roomId)).thenReturn(messages);
-        
-        // when
-        List<ChatMessageDTO> results = chatService.getUnreadMessages(roomId, mentorId);
+        List<ChatMessageDTO> results = chatService.getUnreadMessages(roomId, userId);
         
         // then
         assertNotNull(results);
         assertEquals(1, results.size());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomIdAndReadByMentorFalse(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMenteeFalse(anyString());
+        verify(messageService, times(1)).getUnreadMessages(roomId, userId);
     }
     
     @Test
-    void getUnreadMessages_멘티의_읽지않은_메시지_조회_성공() {
+    void markAsRead_메시지_읽음_처리_성공() {
         // given
         String roomId = "test-room-id";
-        Long menteeId = 2L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        when(messageRepository.findByRoomIdAndReadByMenteeFalse(roomId)).thenReturn(messages);
+        Long userId = 1L;
         
         // when
-        List<ChatMessageDTO> results = chatService.getUnreadMessages(roomId, menteeId);
+        chatService.markAsRead(roomId, userId);
         
         // then
-        assertNotNull(results);
-        assertEquals(1, results.size());
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMentorFalse(anyString());
-        verify(messageRepository, times(1)).findByRoomIdAndReadByMenteeFalse(roomId);
+        verify(messageService, times(1)).markAsRead(roomId, userId);
     }
     
     @Test
-    void getUnreadMessages_권한없는_사용자_예외발생() {
+    void getUnreadMessageCount_읽지않은_메시지_카운트_조회_성공() {
         // given
         String roomId = "test-room-id";
-        Long unauthorizedUserId = 3L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getUnreadMessages(roomId, unauthorizedUserId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMentorFalse(anyString());
-        verify(messageRepository, never()).findByRoomIdAndReadByMenteeFalse(anyString());
-    }
-    
-    @Test
-    void getUnreadMessages_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        Long mentorId = 1L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getUnreadMessages(roomId, mentorId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMentorFalse(anyString());
-        verify(messageRepository, never()).findByRoomIdAndReadByMenteeFalse(anyString());
-    }
-    
-    @Test
-    void markAsRead_멘토_메시지_읽음_처리_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long mentorId = 1L;
-        
-        // 채팅방 설정
-        CoffeeLetterRoom testRoom = new CoffeeLetterRoom();
-        testRoom.setId(roomId);
-        testRoom.setMentorId(mentorId);
-        testRoom.setMenteeId(2L);
-        testRoom.setUnreadCountMentor(3);
-        
-        // 읽지 않은 메시지 설정
-        List<ChatMessage> unreadMessages = Arrays.asList(
-            new ChatMessage(), new ChatMessage(), new ChatMessage()
-        );
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(testRoom));
-        when(messageRepository.findByRoomIdAndReadByMentorFalse(roomId)).thenReturn(unreadMessages);
-        when(messageRepository.saveAll(unreadMessages)).thenReturn(unreadMessages);
-        when(roomRepository.save(testRoom)).thenReturn(testRoom);
-        
-        // when
-        chatService.markAsRead(roomId, mentorId);
-        
-        // then
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomIdAndReadByMentorFalse(roomId);
-        verify(messageRepository, times(1)).saveAll(unreadMessages);
-        verify(roomRepository, times(1)).save(testRoom);
-        
-        // 메시지들이 모두 읽음 처리되었는지 확인
-        for (ChatMessage message : unreadMessages) {
-            assertTrue(message.isReadByMentor());
-        }
-        
-        // 읽지 않은 메시지 카운트가 0으로 초기화되었는지 확인
-        assertEquals(0, testRoom.getUnreadCountMentor());
-    }
-    
-    @Test
-    void markAsRead_멘티_메시지_읽음_처리_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long menteeId = 2L;
-        
-        // 채팅방 설정
-        CoffeeLetterRoom testRoom = new CoffeeLetterRoom();
-        testRoom.setId(roomId);
-        testRoom.setMentorId(1L);
-        testRoom.setMenteeId(menteeId);
-        testRoom.setUnreadCountMentee(2);
-        
-        // 읽지 않은 메시지 설정
-        List<ChatMessage> unreadMessages = Arrays.asList(
-            new ChatMessage(), new ChatMessage()
-        );
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(testRoom));
-        when(messageRepository.findByRoomIdAndReadByMenteeFalse(roomId)).thenReturn(unreadMessages);
-        when(messageRepository.saveAll(unreadMessages)).thenReturn(unreadMessages);
-        when(roomRepository.save(testRoom)).thenReturn(testRoom);
-        
-        // when
-        chatService.markAsRead(roomId, menteeId);
-        
-        // then
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, times(1)).findByRoomIdAndReadByMenteeFalse(roomId);
-        verify(messageRepository, times(1)).saveAll(unreadMessages);
-        verify(roomRepository, times(1)).save(testRoom);
-        
-        // 메시지들이 모두 읽음 처리되었는지 확인
-        for (ChatMessage message : unreadMessages) {
-            assertTrue(message.isReadByMentee());
-        }
-        
-        // 읽지 않은 메시지 카운트가 0으로 초기화되었는지 확인
-        assertEquals(0, testRoom.getUnreadCountMentee());
-    }
-    
-    @Test
-    void markAsRead_권한없는_사용자_예외발생() {
-        // given
-        String roomId = "test-room-id";
-        Long unauthorizedUserId = 3L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.markAsRead(roomId, unauthorizedUserId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMentorFalse(anyString());
-        verify(messageRepository, never()).findByRoomIdAndReadByMenteeFalse(anyString());
-        verify(messageRepository, never()).saveAll(any());
-    }
-    
-    @Test
-    void markAsRead_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        Long mentorId = 1L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.markAsRead(roomId, mentorId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-        verify(messageRepository, never()).findByRoomIdAndReadByMentorFalse(anyString());
-        verify(messageRepository, never()).findByRoomIdAndReadByMenteeFalse(anyString());
-        verify(messageRepository, never()).saveAll(any());
-    }
-    
-    @Test
-    void getUnreadMessageCount_멘토_카운트_조회_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long mentorId = 1L;
+        Long userId = 1L;
         int expectedCount = 3;
         
-        CoffeeLetterRoom testRoom = new CoffeeLetterRoom();
-        testRoom.setId(roomId);
-        testRoom.setMentorId(mentorId);
-        testRoom.setMenteeId(2L);
-        testRoom.setUnreadCountMentor(expectedCount);
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(testRoom));
+        when(messageService.getUnreadMessageCount(roomId, userId)).thenReturn(expectedCount);
         
         // when
-        int actualCount = chatService.getUnreadMessageCount(roomId, mentorId);
+        int actualCount = chatService.getUnreadMessageCount(roomId, userId);
         
         // then
         assertEquals(expectedCount, actualCount);
-        verify(roomRepository, times(1)).findById(roomId);
-    }
-    
-    @Test
-    void getUnreadMessageCount_멘티_카운트_조회_성공() {
-        // given
-        String roomId = "test-room-id";
-        Long menteeId = 2L;
-        int expectedCount = 5;
-        
-        CoffeeLetterRoom testRoom = new CoffeeLetterRoom();
-        testRoom.setId(roomId);
-        testRoom.setMentorId(1L);
-        testRoom.setMenteeId(menteeId);
-        testRoom.setUnreadCountMentee(expectedCount);
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(testRoom));
-        
-        // when
-        int actualCount = chatService.getUnreadMessageCount(roomId, menteeId);
-        
-        // then
-        assertEquals(expectedCount, actualCount);
-        verify(roomRepository, times(1)).findById(roomId);
-    }
-    
-    @Test
-    void getUnreadMessageCount_권한없는_사용자_예외발생() {
-        // given
-        String roomId = "test-room-id";
-        Long unauthorizedUserId = 3L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.of(room));
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getUnreadMessageCount(roomId, unauthorizedUserId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
-    }
-    
-    @Test
-    void getUnreadMessageCount_존재하지_않는_채팅방_예외발생() {
-        // given
-        String roomId = "non-existent-room-id";
-        Long mentorId = 1L;
-        
-        when(roomRepository.findById(roomId)).thenReturn(Optional.empty());
-        
-        // when & then
-        assertThrows(IllegalArgumentException.class, () -> {
-            chatService.getUnreadMessageCount(roomId, mentorId);
-        });
-        
-        verify(roomRepository, times(1)).findById(roomId);
+        verify(messageService, times(1)).getUnreadMessageCount(roomId, userId);
     }
     
     @Test
     void getLastMessage_성공적인_조회() {
         // given
         String roomId = "test-room-id";
-        LocalDateTime timestamp = LocalDateTime.now();
         
-        ChatMessage lastMessage = new ChatMessage();
-        lastMessage.setId("last-message-id");
-        lastMessage.setRoomId(roomId);
-        lastMessage.setSenderId(2L);
-        lastMessage.setSenderName("멘티");
-        lastMessage.setContent("마지막 메시지입니다");
-        lastMessage.setTimestamp(timestamp);
-        lastMessage.setType(MessageType.CHAT);
-        lastMessage.setReadByMentor(false);
-        lastMessage.setReadByMentee(true);
-        
-        ChatMessageDTO expectedDTO = new ChatMessageDTO();
-        expectedDTO.setId("last-message-id");
-        expectedDTO.setRoomId(roomId);
-        expectedDTO.setSenderId(2L);
-        expectedDTO.setSenderName("멘티");
-        expectedDTO.setContent("마지막 메시지입니다");
-        expectedDTO.setTimestamp(timestamp);
-        expectedDTO.setType(MessageType.CHAT);
-        expectedDTO.setReadByMentor(false);
-        expectedDTO.setReadByMentee(true);
-        
-        when(messageRepository.findFirstByRoomIdOrderByTimestampDesc(roomId)).thenReturn(lastMessage);
-        when(modelMapper.map(lastMessage, ChatMessageDTO.class)).thenReturn(expectedDTO);
+        when(messageService.getLastMessage(roomId)).thenReturn(messageDTO);
         
         // when
         ChatMessageDTO result = chatService.getLastMessage(roomId);
         
         // then
         assertNotNull(result);
-        assertEquals(expectedDTO.getId(), result.getId());
-        assertEquals(expectedDTO.getRoomId(), result.getRoomId());
-        assertEquals(expectedDTO.getSenderId(), result.getSenderId());
-        assertEquals(expectedDTO.getSenderName(), result.getSenderName());
-        assertEquals(expectedDTO.getContent(), result.getContent());
-        assertEquals(expectedDTO.getType(), result.getType());
-        assertEquals(expectedDTO.getTimestamp(), result.getTimestamp());
-        assertEquals(expectedDTO.isReadByMentor(), result.isReadByMentor());
-        assertEquals(expectedDTO.isReadByMentee(), result.isReadByMentee());
-        
-        verify(messageRepository, times(1)).findFirstByRoomIdOrderByTimestampDesc(roomId);
-        verify(modelMapper, times(1)).map(lastMessage, ChatMessageDTO.class);
+        assertEquals(messageDTO.getId(), result.getId());
+        assertEquals(messageDTO.getRoomId(), result.getRoomId());
+        verify(messageService, times(1)).getLastMessage(roomId);
     }
     
     @Test
     void getLastMessage_메시지가_없는_경우() {
         // given
         String roomId = "empty-room-id";
-        when(messageRepository.findFirstByRoomIdOrderByTimestampDesc(roomId)).thenReturn(null);
+        when(messageService.getLastMessage(roomId)).thenReturn(null);
         
         // when
         ChatMessageDTO result = chatService.getLastMessage(roomId);
         
         // then
         assertNull(result);
-        verify(messageRepository, times(1)).findFirstByRoomIdOrderByTimestampDesc(roomId);
-        verify(modelMapper, times(0)).map(any(), any());
-    }
-    
-    @Test
-    void getLastMessage_시스템_메시지_조회() {
-        // given
-        String roomId = "test-room-id";
-        LocalDateTime timestamp = LocalDateTime.now();
-        
-        ChatMessage systemMessage = new ChatMessage();
-        systemMessage.setId("system-message-id");
-        systemMessage.setRoomId(roomId);
-        systemMessage.setSenderId(0L);
-        systemMessage.setSenderName("System");
-        systemMessage.setContent("채팅방이 개설되었습니다");
-        systemMessage.setTimestamp(timestamp);
-        systemMessage.setType(MessageType.SYSTEM);
-        systemMessage.setReadByMentor(true);
-        systemMessage.setReadByMentee(true);
-        
-        ChatMessageDTO expectedDTO = new ChatMessageDTO();
-        expectedDTO.setId("system-message-id");
-        expectedDTO.setRoomId(roomId);
-        expectedDTO.setSenderId(0L);
-        expectedDTO.setSenderName("System");
-        expectedDTO.setContent("채팅방이 개설되었습니다");
-        expectedDTO.setTimestamp(timestamp);
-        expectedDTO.setType(MessageType.SYSTEM);
-        expectedDTO.setReadByMentor(true);
-        expectedDTO.setReadByMentee(true);
-        
-        when(messageRepository.findFirstByRoomIdOrderByTimestampDesc(roomId)).thenReturn(systemMessage);
-        when(modelMapper.map(systemMessage, ChatMessageDTO.class)).thenReturn(expectedDTO);
-        
-        // when
-        ChatMessageDTO result = chatService.getLastMessage(roomId);
-        
-        // then
-        assertNotNull(result);
-        assertEquals(expectedDTO.getId(), result.getId());
-        assertEquals(expectedDTO.getRoomId(), result.getRoomId());
-        assertEquals(expectedDTO.getSenderId(), result.getSenderId());
-        assertEquals(expectedDTO.getSenderName(), result.getSenderName());
-        assertEquals(expectedDTO.getContent(), result.getContent());
-        assertEquals(expectedDTO.getType(), result.getType());
-        assertEquals(MessageType.SYSTEM, result.getType());
-        
-        verify(messageRepository, times(1)).findFirstByRoomIdOrderByTimestampDesc(roomId);
-        verify(modelMapper, times(1)).map(systemMessage, ChatMessageDTO.class);
+        verify(messageService, times(1)).getLastMessage(roomId);
     }
 }
